Conflicts:
	src/qml/qml/qqmlbinding.cpp
	src/qml/jsruntime/qv4arraybuffer.cpp
	src/qml/jsruntime/qv4functionobject.cpp

+++ src/quick/scenegraph/qsgcontext.cpp
@@ -256,12 +256,12 @@ public:
         advanceAnimation();
     }
 
-    float m_time;
-    float m_vsync;
+    double m_time;
+    double m_vsync;
     Mode m_mode;
     QElapsedTimer m_timer;
     QElapsedTimer m_wallTime;
-    float m_lag;
+    double m_lag;
     int m_bad;
     int m_good;
 };
